detail biodata

+++ application/models/Biodata_model.php
@@ -12,6 +12,12 @@ class Biodata_model extends CI_Model {
 
 	public function getBiodata()
 	{
+		$this->db->order_by('full_name', 'asc');
+		return $this->db->get('biodata')->result_array();
+	}
+
+	public function getBiodataJoinUser()
+	{
 		// full outer join
 		// $query = "SELECT * FROM biodata
 		// 		  LEFT JOIN users ON biodata.id_biodata = users.id_biodata
@@ -28,19 +34,26 @@ class Biodata_model extends CI_Model {
 		return $this->db->get_where('biodata', ['id_biodata' => $id])->row_array();
 	}
 
+	public function getBiodataJoinById($id)
+	{
+		$this->db->join('users', 'users.id_biodata = biodata.id_biodata', 'left');
+		$this->db->join('roles', 'roles.id_role = users.id_role', 'left');
+		return $this->db->get_where('biodata', ['biodata.id_biodata' => $id])->row_array();
+	}
+
 	public function insertBiodata()
 	{
 		$dataUser = $this->mamo->dataUser();
 
 		$data = [
-			'full_name' => $this->input->post('full_name', true),
+			'full_name' => ucwords(strtolower($this->input->post('full_name', true))),
 			'phone_number' => $this->input->post('phone_number', true),
-			'email' => $this->input->post('email', true),
+			'email' => strtolower($this->input->post('email', true)),
 			'address' => $this->input->post('address', true),
 			'gender' => $this->input->post('gender', true)
 		];
 		
-		$this->mamo->checkAccessAdministratorOnly("menambahkan Biodata " . $data['full_name'] . 
+		$this->mamo->checkAccessAdministratorOnly("menambahkan Karyawan " . $data['full_name'] . 
 			" No. Telepon "   . $data['phone_number'] . 
 			" Email "  		  . $data['email'] . 
 			" Alamat " 		  . $data['address'] . 
@@ -66,7 +79,7 @@ class Biodata_model extends CI_Model {
 			$config['upload_path'] = './assets/file/file_cv/';
 			$config['allowed_types'] = 'doc|docx|pdf|xls|xlsx';
 		
-			$this->load->library('upload', $config);
+			$this->upload->initialize($config);
 		
 			if ($this->upload->do_upload('file_cv')) {
 				$new_file_cv = $this->upload->data('file_name');
@@ -78,8 +91,8 @@ class Biodata_model extends CI_Model {
 		
 
 		$this->db->insert('biodata', $data);
-		$this->session->set_flashdata('message-success', 'Biodata ' . $data['full_name'] . ' berhasil ditambahkan');
-		$this->lomo->insertLog($dataUser['username'] . " berhasil menambahkan Biodata " . $data['full_name'] . 
+		$this->session->set_flashdata('message-success', 'Karyawan ' . $data['full_name'] . ' berhasil ditambahkan');
+		$this->lomo->insertLog($dataUser['username'] . " berhasil menambahkan Karyawan " . $data['full_name'] . 
 			" No. Telepon "   . $data['phone_number'] . 
 			" Email "   	  . $data['email'] . 
 			" Alamat " 		  . $data['address'] . 
@@ -97,8 +110,9 @@ class Biodata_model extends CI_Model {
 		$email = $db['email'];
 		$address = $db['address'];
 		$gender = $db['gender'];
-
-		$this->mamo->checkAccessAdministratorOnly("mengubah Biodata " . $full_name);
+		if ($db['id_biodata'] != $dataUser['id_biodata']) {
+			$this->mamo->checkAccessAdministratorOnly("mengubah Karyawan " . $full_name);
+		}
 
 		$photo = $_FILES['photo']['name'];
 		if ($photo) {
@@ -139,17 +153,17 @@ class Biodata_model extends CI_Model {
 		}
 		
 		$data = [
-			'full_name' => $this->input->post('full_name', true),
+			'full_name' => ucwords(strtolower($this->input->post('full_name', true))),
 			'phone_number' => $this->input->post('phone_number', true),
-			'email' => $this->input->post('email', true),
+			'email' => strtolower($this->input->post('email', true)),
 			'address' => $this->input->post('address', true),
 			'gender' => $this->input->post('gender', true)
 		];
 
 
 		$this->db->update('biodata', $data, ['id_biodata' => $id]);
-		$this->session->set_flashdata('message-success', 'Biodata ' . $full_name . ' berhasil diubah');
-		$this->lomo->insertLog($dataUser['username'] . " berhasil mengubah Biodata " . 
+		$this->session->set_flashdata('message-success', 'Karyawan ' . $full_name . ' berhasil diubah');
+		$this->lomo->insertLog($dataUser['username'] . " berhasil mengubah Karyawan " . 
 			$full_name . " menjadi " . $data['full_name'] . ", " .
 			" mengubah No. Telepon " . $phone_number. " menjadi " . $data['phone_number'] . ", " .
 			" mengubah Email " . $email. " menjadi " . $data['email'] . ", " .
@@ -167,13 +181,15 @@ class Biodata_model extends CI_Model {
 		$db = $this->getBiodataById($id);
 		$full_name = $db['full_name'];
 		
-		$this->mamo->checkAccessAdministratorOnly("menghapus Biodata " . $full_name);
-
-		unlink(FCPATH . 'assets/img/img_profiles/' . $db['photo']);
+		$this->mamo->checkAccessAdministratorOnly("menghapus Karyawan " . $full_name);
+		$old_photo = $db['photo'];
+		if ($old_photo != 'default.png') {
+			unlink(FCPATH . 'assets/img/img_profiles/' . $db['photo']);
+		}
 		unlink(FCPATH . 'assets/file/file_cv/' . $db['file_cv']);
 		$this->db->delete('biodata', ['id_biodata' => $id]);
-		$this->session->set_flashdata('message-success', 'Biodata ' . $full_name . ' berhasil dihapus');
-		$this->lomo->insertLog($dataUser['username'] . " berhasil menghapus Biodata " . $full_name, $dataUser['id_user']);
+		$this->session->set_flashdata('message-success', 'Karyawan ' . $full_name . ' berhasil dihapus');
+		$this->lomo->insertLog($dataUser['username'] . " berhasil menghapus Karyawan " . $full_name, $dataUser['id_user']);
 		redirect('biodata');
 	}
 }
